API OMDB/requirements.txt
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.0.1
fonttools==4.55.0
idna==3.10
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.2.2
polars==1.17.1
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
pyzmq==26.0.3
requests==2.32.3
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
xlrd==2.0.1
